+++ common/include/pcl/common/impl/io.hpp
@@ -159,7 +159,7 @@ pcl::copyPointCloud (const pcl::PointCloud<PointT> &cloud_in,
   // Allocate enough space and copy the basics
   cloud_out.points.resize (indices.size ());
   cloud_out.header   = cloud_in.header;
-  cloud_out.width    = indices.size ();
+  cloud_out.width    = static_cast<uint32_t> (indices.size ());
   cloud_out.height   = 1;
   if (cloud_in.is_dense)
     cloud_out.is_dense = true;
